% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound}
\alias{bound}
\title{Bound a vector}
\usage{
bound(x, min = 0.01, max = 0.99)
}
\arguments{
\item{x}{Numeric vector to be bounded}

\item{min}{Minimum allowed value for vector "x"; default is \code{0.01}}

\item{max}{Maximum allowed value for vector "x"; default is \code{0.99}}
}
\value{
Numeric vector of the same length as \code{x} with no values less than \code{minimum} nor greater than \code{maximum}.
}
\description{
This unexported helper function bounds a numeric vector on a minimum and maximum value.
}
