% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_counts}
\alias{expand_counts}
\title{Expand aggregated count data to observation-level}
\usage{
expand_counts(.data, freq_col)
}
\arguments{
\item{.data}{data frame}

\item{freq_col}{column containing the frequency values}
}
\value{
tibble
}
\description{
Expand aggregated count data to observation-level
}
