% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{fisher}
\alias{fisher}
\title{Fisher test}
\usage{
fisher(df, exposure, outcome, ...)
}
\arguments{
\item{df}{dataframe}

\item{exposure}{exposure}

\item{outcome}{outcome}

\item{...}{additional arguments passed to twoxtwo function}
}
\value{
tibble
}
\description{
Fisher test
}
