% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{odds_ratio}
\alias{odds_ratio}
\title{Calculate odds ratio}
\usage{
odds_ratio(.data, exposure, outcome, alpha = 0.05, ...)
}
\arguments{
\item{.data}{dataframe}

\item{exposure}{exposure}

\item{outcome}{outcome}

\item{alpha}{significance level}

\item{...}{additional arguments passed to \code{\link[=twoxtwo]{twoxtwo()}}}
}
\value{
\code{tibble}
}
\description{
Calculate odds ratio
}
\seealso{
Other Effect measures: 
\code{\link{risk_diff}()},
\code{\link{risk_ratio}()}
}
\concept{Effect measures}
