% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.twoxtwo}
\alias{print.twoxtwo}
\title{Print twoxtwo object}
\usage{
\method{print}{twoxtwo}(x, ...)
}
\arguments{
\item{x}{\link[twoxtwo]{twoxtwo} object}

\item{...}{Additional arguments passed to \link[knitr]{kable}}
}
\value{
print output ...
}
\description{
description here ...
}
