% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{reri}
\alias{reri}
\title{Relative risk due to interaction (RERI)}
\usage{
reri(.data, exposure, outcome, effect_modifier, ...)
}
\arguments{
\item{.data}{Data frame with observation-level exposure and outcome data}

\item{exposure}{Name of exposure variable}

\item{outcome}{Name of outcome variable}

\item{effect_modifier}{Name of effect modifier variable}

\item{...}{Additional arguments passed to \link[twoxtwo]{twoxtwo} function}
}
\value{
\code{tibble}
}
\description{
Relative risk due to interaction (RERI)
}
