% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{risk_ratio}
\alias{risk_ratio}
\title{Calculate risk ratio}
\usage{
risk_ratio(df, exposure, outcome, alpha = 0.05, ...)
}
\arguments{
\item{df}{dataframe}

\item{exposure}{exposure}

\item{outcome}{outcome}

\item{alpha}{significance level}

\item{...}{additional arguments passed to twoxtwo function}
}
\value{
tibble
}
\description{
Calculate risk ratio
}
