% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Expanded Titanic dataset}
\format{
A data frame with 2201 rows and 4 variables:
\itemize{
\item \strong{Class}: Passenger class ("1st", "2nd", "3rd") or crew status ("Crew")
\item \strong{Sex}: Sex of individual ("Male" or "Female")
\item \strong{Age}: Categorized age ("Adult" or "Child")
\item \strong{Survived}: Whether or not individual survived ("Yes" or "No")
}
}
\usage{
titanic
}
\description{
This data is based on the \link[datasets]{Titanic} dataset. Unlike the version in the \code{datasets} package, the data here is expanded to the observation-level rather than cross-tabulated.
}
\examples{
head(titanic)
}
\keyword{datasets}
