% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display}
\alias{display}
\title{Title}
\usage{
display(.twoxtwo, ...)
}
\arguments{
\item{.twoxtwo}{\link[twoxtwo]{twoxtwo} object}

\item{...}{Additional arguments passed to \link[knitr]{kable}}
}
\value{
kable output ...
}
\description{
Title
}
