% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{fisher}
\alias{fisher}
\title{Fisher test}
\usage{
fisher(.data, exposure, outcome, ...)
}
\arguments{
\item{.data}{Data frame with observation-level exposure and outcome data}

\item{exposure}{Name of exposure variable}

\item{outcome}{Name of outcome variable}

\item{...}{Additional arguments passed to \link[twoxtwo]{twoxtwo} function}
}
\value{
tibble
}
\description{

}
