% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_measure}
\alias{format_measure}
\title{Title}
\usage{
format_measure(.data, digits = 3)
}
\arguments{
\item{.data}{Output from a twoxtwo effect measure function (e.g. \link[twoxtwo]{odds_ratio})}

\item{digits}{Number of digits; default is \code{3}}
}
\value{
formatted measure
}
\description{
Title
}
