% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{measures}
\alias{measures}
\alias{odds_ratio}
\alias{risk_ratio}
\alias{risk_diff}
\title{Effect measures}
\usage{
odds_ratio(.data, exposure, outcome, alpha = 0.05, ...)

risk_ratio(.data, exposure, outcome, alpha = 0.05, ...)

risk_diff(.data, exposure, outcome, alpha = 0.05, ...)
}
\arguments{
\item{.data}{Data frame with observation-level exposure and outcome data}

\item{exposure}{Name of exposure variable}

\item{outcome}{Name of outcome variable}

\item{alpha}{Significance level to be used for constructing confidence interval; default is \code{0.05}}

\item{...}{Additional arguments passed to \link[twoxtwo]{twoxtwo} function}
}
\value{
\code{tibble}
}
\description{
\subsection{odds ratio}{

lorem ipsum
}

\subsection{risk ratio}{

lorem ipsum
}

\subsection{risk difference}{

lorem ipsum
}
}
