% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{odds_ratio}
\alias{odds_ratio}
\title{Calculate odds ratio}
\usage{
odds_ratio(df, exposure, outcome, alpha = 0.05, ...)
}
\arguments{
\item{df}{dataframe}

\item{exposure}{exposure}

\item{outcome}{outcome}

\item{alpha}{significance level}

\item{...}{additional arguments passed to twoxtwo function}
}
\value{
tibble
}
\description{
Calculate odds ratio
}
