% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.twoxtwo}
\alias{print.twoxtwo}
\title{Print twoxtwo object}
\usage{
\method{print}{twoxtwo}(.twoxtwo, ...)
}
\arguments{
\item{.twoxtwo}{\link[twoxtwo]{twoxtwo} object}

\item{...}{Additional arguments passed to \link[knitr]{kable}}
}
\value{

}
\description{
Print twoxtwo object
}
