% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.twoxtwo}
\alias{print.twoxtwo}
\title{Print twoxtwo object}
\usage{
\method{print}{twoxtwo}(x, ...)
}
\arguments{
\item{x}{\link[twoxtwo]{twoxtwo} object}

\item{...}{Additional arguments passed to \link[knitr]{kable}}
}
\value{
A printed \code{knitr_kable} object with the \code{twoxtwo} cell counts, exposure levels as row names, and outcome levels as column names.
}
\description{
The \code{print.twoxtwo()} function provides an S3 method for printing objects created with \link[twoxtwo]{twoxtwo}. The printed output formats the contents of the \code{twoxtwo} table as a \link[knitr]{kable}.
}
