% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{reri}
\alias{reri}
\title{Relative risk due to interaction (RERI)}
\usage{
reri(.data, exposure, outcome, effect_modifier, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Relative risk due to interaction (RERI)
}
