% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{reri}
\alias{reri}
\title{Relative excess risk due to interaction (RERI)}
\usage{
reri(.data, exposure, outcome, effect_modifier, ...)
}
\arguments{
\item{.data}{Data frame with observation-level exposure and outcome data}

\item{exposure}{Name of exposure variable}

\item{outcome}{Name of outcome variable}

\item{effect_modifier}{Name of effect modifier variable}

\item{...}{Additional arguments passed to \link[twoxtwo]{twoxtwo} function}
}
\value{
A \code{tibble} with the following columns
\itemize{
\item \strong{estimate}: Value for the RERI based on exposure, outcome, and effect modifier variables specified
\item \strong{exposure}: Name of the exposure variable followed by +/- levels (e.g. smoking::yes/no)
\item \strong{outcome}: Name of the outcome variable followed by +/- levels (e.g. heart_disease::yes/no)
\item \strong{effect_modifier}: Name of the effect modifier followed by levels (e.g. diabetes::TRUE/FALSE)
}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}]

This function computes the RERI, which is a metric for assessing additive interaction in relative risk. The input data is first stratified by the effect modifier variable. The function then internally uses \link[twoxtwo]{twoxtwo} to create the two-by-two contingency table and compute risk within strata. The RERI is then calculatated from these stratum-specific risks.

The formulas used to calculate the RERI is available in 'Details'.
}
\details{
\deqn{RERI = (r11/r00) - (r10/r00) - (r01/r00) + 1}
}
\references{
Richardson, D. B., & Kaufman, J. S. (2009). Estimation of the relative excess risk due to interaction and associated confidence bounds. American journal of epidemiology, 169(6), 756–760. https://doi.org/10.1093/aje/kwn411
}
