% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{risk_ratio}
\alias{risk_ratio}
\title{Calculate risk ratio}
\usage{
risk_ratio(.data, exposure, outcome, alpha = 0.05, ...)
}
\arguments{
\item{.data}{dataframe}

\item{exposure}{exposure}

\item{outcome}{outcome}

\item{alpha}{significance level}

\item{...}{additional arguments passed to \code{\link[=twoxtwo]{twoxtwo()}}}
}
\value{
\code{tibble}
}
\description{
Calculate risk ratio
}
\seealso{
Other Effect measures: 
\code{\link{odds_ratio}()},
\code{\link{risk_diff}()}
}
\concept{Effect measures}
