% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.twoxtwo}
\alias{summary.twoxtwo}
\title{Summarize twoxtwo object}
\usage{
\method{summary}{twoxtwo}(.twoxtwo, alpha = 0.05, ...)
}
\arguments{
\item{.twoxtwo}{\link[twoxtwo]{twoxtwo} object}

\item{alpha}{Sigificance level to be used for constructing confidence interval; default is \code{0.05}}

\item{...}{Additional arguments passed to \link[twoxtwo]{print.twoxtwo}}
}
\value{

}
\description{
Summarize twoxtwo object
}
