% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoxtwo.R
\name{twoxtwo}
\alias{twoxtwo}
\title{Create a two-by-two table}
\usage{
twoxtwo(.data, exposure, outcome, levels = NULL, na.rm = TRUE, verbose = FALSE)
}
\arguments{
\item{.data}{Data frame with observation-level exposure and outcome data}

\item{exposure}{Name of exposure variable}

\item{outcome}{Name of outcome variable}

\item{levels}{explicitly specify levels for the exposure and outcome as a named list; if supplied, contingency table will be oriented with respect to the sequence of levels specified}

\item{na.rm}{logical as to whether or not to remove NA values when constructing contingency table; default is TRUE}

\item{verbose}{Logical}
}
\value{
tibble
}
\description{
Create a two-by-two table
}
