% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoxtwo.R
\name{twoxtwo}
\alias{twoxtwo}
\title{Create a two-by-two table}
\usage{
twoxtwo(df, exposure, outcome, levels = NULL, na.rm = TRUE)
}
\arguments{
\item{df}{data frame}

\item{exposure}{exposure}

\item{outcome}{outcome}

\item{levels}{explicitly specify levels for the exposure and outcome as a named list; if supplied, contingency table will be oriented with respect to the sequence of levels specified}

\item{na.rm}{logical as to whether or not to remove NA values when constructing contingency table; default is TRUE}
}
\value{
tibble
}
\description{
Create a two-by-two table
}
